requirements.txt

# certifi                  2022.9.24
click                    8.1.3
Flask                    2.2.2
future                   0.18.2
idna                     3.4
itsdangerous             2.1.2
Jinja2                   3.1.2
line-bot-sdk             2.3.0
MarkupSafe               2.1.1
graphviz                 0.20.1
python-dotenv            0.21.0
requests                 2.28.1
transitions              0.9.0
urllib3                  1.26.12
Werkzeug                 2.2.2
